seperate home page and main
